deplacer la fonciton anonymizedIdentite vers anonymiserData.js
