+++ backend/src/util/anonymiserData.js
@@ -1,3 +1,5 @@
+import { TYPE_IMMATRICULATION } from '../constant/type.js'
+
 export const anonymize = (text, nbVisibleCharAtPrefixAndSuffix = 2) => {
   const anonymizedText = '*'.repeat(text.length - nbVisibleCharAtPrefixAndSuffix * 2)
   return text.substr(0, nbVisibleCharAtPrefixAndSuffix) + anonymizedText + text.substr(nbVisibleCharAtPrefixAndSuffix + anonymizedText.length)
@@ -36,5 +38,25 @@ export const anonymizedControlesTechniques = (controlesTechniques) => {
   return {
     ct: anonymizedCt,
     ctUpdateDate,
+  }
+}
+
+export const anonymizedIdentite = (identiteLog) => {
+  const anonymizedDataIdentite =
+    {
+      alreadyHasIdAndKey: identiteLog.alreadyHasIdAndKey,
+      uuid: identiteLog.uuid,
+      anonymizedNom: identiteLog.nom ? anonymizeIdentite(identiteLog.nom) : undefined,
+      anonymizedPrenom: (identiteLog.prenoms && identiteLog.typeImmatriculation === TYPE_IMMATRICULATION.SIV) ? anonymizeIdentite(identiteLog.prenoms[0]) : undefined,
+      anonymizedRaisonSociale: identiteLog.raisonSociale ? anonymizeIdentite(identiteLog.raisonSociale) : undefined,
+      anonymizedSiren: identiteLog.siren ? anonymizeIdentite(identiteLog.siren) : undefined,
+      anonymizedNumeroImmatriculation: identiteLog.numeroImmatriculation ? anonymizeIdentite(identiteLog.numeroImmatriculation) : undefined,
+      anonymizedNumeroFormule: identiteLog.numeroFormule ? anonymizeIdentite(identiteLog.numeroFormule) : undefined,
+      dateEmissionCertificatImmatriculation: identiteLog.dateEmissionCertificatImmatriculation,
+      id: identiteLog.id,
+      key: identiteLog.key,
+    }
+  return {
+    anonymizedDataIdentite,
   }
 }
